Refactored using a Notifier per feature.
Added configuration string validations in RedisMessageBusBroker and RedisOutputCacheStorageProvider
