+++ src/Orchard.Web/Modules/Orchard.Redis/Orchard.Redis.csproj
@@ -81,7 +81,9 @@
   </ItemGroup>
   <ItemGroup>
     <Content Include="Web.config" />
-    <Compile Include="Configuration\RedisNotificationProvider.cs" />
+    <Compile Include="Services\RedisMessageBusNotificationProvider.cs" />
+    <Compile Include="Services\RedisOutputCacheNotificationProvider.cs" />
+    <Compile Include="Services\RedisCacheNotificationProvider.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Content Include="Module.txt" />
   </ItemGroup>
