+++ src/Orchard.Web/Modules/Orchard.Redis/OutputCache/RedisOutputCacheStorageProvider.cs
@@ -74,14 +74,26 @@ namespace Orchard.Redis.OutputCache {
         }
 
         public void Remove(string key) {
+            if (_connectionMultiplexer == null) {
+                return;
+            }
+
             Database.KeyDelete(GetLocalizedKey(key));
         }
 
         public void RemoveAll() {
+            if (_connectionMultiplexer == null) {
+                return;
+            }
+
             Database.KeyDeleteWithPrefix(GetLocalizedKey("*"));
         }
 
         public CacheItem GetCacheItem(string key) {
+            if (_connectionMultiplexer == null) {
+                return null;
+            }
+
             var value = Database.StringGet(GetLocalizedKey(key));
 
             if (value.IsNullOrEmpty) {
@@ -110,6 +122,10 @@ namespace Orchard.Redis.OutputCache {
         }
 
         public int GetCacheItemsCount() {
+            if (_connectionMultiplexer == null) {
+                return 0;
+            }
+
             return Database.KeyCount(GetLocalizedKey("*"));
         }
 
@@ -127,6 +143,10 @@ namespace Orchard.Redis.OutputCache {
         /// </summary>
         /// <returns>The keys for the current tenant.</returns>
         private IEnumerable<string> GetAllKeys() {
+            if (_connectionMultiplexer == null) {
+                return new string[0];
+            }
+
             // prevent the same request from computing the list twice (count + list)
             if (_keysCache == null) {
                 _keysCache = new HashSet<string>();
